+++ boot/pod/src/boot/from/io/aviso/repl.clj
@@ -1,72 +1,65 @@
 (ns boot.from.io.aviso.repl
-  {:boot/from :AvisoNovate/pretty
-   :doc "Utilities to assist with REPL-oriented development"}
-  (:use
-    [boot.from.io.aviso.exception])
+  "Utilities to assist with REPL-oriented development"
+  {:boot/from :AvisoNovate/pretty:0.1.33}
   (:require
-    [clojure
-     [main :as main]
-     [repl :as repl]
-     [stacktrace :as st]]))
-
-(defn standard-frame-filter
-  "Default stack frame filter used when printing REPL exceptions. This will omit frames in the `clojure.lang`
-  and `java.lang.reflect` package, hide frames in the `sun.reflect` package,
-  and terminates the stack trace at the read-eval-print loop frame. This tends to be very concise; you can use
-  `(io.aviso.exception/write-exception *e)` to display the full stack trace without filtering."
-  [frame]
-  (let [package (-> frame :package str)]
-    (cond
-      (= package "clojure.lang")
-      :omit
-
-      (.startsWith package "sun.reflect")
-      :hide
-
-      (= package "java.lang.reflect")
-      :omit
-
-      (= package "org.projectodd.shimdandy.impl")
-      :omit
-      
-      (.startsWith package "java.")
-      :omit
-      
-      (.startsWith ^String (:name frame) "clojure.main/repl/read-eval-print")
-      :terminate
-
-      :else
-      :show)))
+    [boot.from.io.aviso.exception :as e]
+    [clojure.pprint :refer [pprint write]]
+    [clojure.main :as main]
+    [clojure.repl :as repl]
+    [clojure.stacktrace :as st]
+    [clojure.edn :as edn]
+    [boot.from.io.aviso.writer :as writer])
+  (:import
+    (clojure.lang RT)))
 
 (defn- reset-var!
   [v override]
   (alter-var-root v (constantly override)))
 
-(defn- write
+(defn- print-exception
   [e options]
-  (print (format-exception e (assoc options :filter standard-frame-filter)))
+  (print (e/format-exception e options))
   (flush))
 
 (defn pretty-repl-caught
   "A replacement for `clojure.main/repl-caught` that prints the exception to `*err*`, without a stack trace or properties."
   [e]
-  (write e {:frame-limit 0 :properties false}))
+  (print-exception e {:frame-limit 0 :properties false}))
+
+(defn uncaught-exception-handler
+  "Creates a reified UncaughtExceptionHandler that prints the formatted exception to `*err*`."
+  {:added "0.1.18"}
+  []
+  (reify Thread$UncaughtExceptionHandler
+    (uncaughtException [_ _ t]
+      (binding [*out* *err*]
+        (printf "Uncaught exception in thread %s:%n" (-> (Thread/currentThread) .getName))
+        (e/write-exception t)))))
+
 
 (defn pretty-pst
-  "Used as an override of `clojure.repl/pst` but uses pretty formatting. The optional parameter must be an exception
-  (it can not be a depth, as with the standard implementation of `pst`)."
+  "Used as an override of `clojure.repl/pst` but uses pretty formatting."
   ([] (pretty-pst *e))
-  ([e] (write e nil)))
+  ([e-or-depth]
+   (if (instance? Throwable e-or-depth)
+     (print-exception e-or-depth nil)
+     (pretty-pst *e e-or-depth)))
+  ([e depth] (print-exception e {:frame-limit depth})))
 
 (defn pretty-print-stack-trace
-  "Replacement for `clojure.stracktrace/print-stack-trace` and `print-cause-trace`. These functions are used by `clojure.test`."
+  "Replacement for `clojure.stacktrace/print-stack-trace` and `print-cause-trace`. These functions are used by `clojure.test`."
   ([tr] (pretty-print-stack-trace tr nil))
   ([tr n]
-   (write tr {:frame-limit n})))
+   (println)
+   (print-exception tr {:frame-limit n})))
 
 (defn install-pretty-exceptions
   "Installs an override that outputs pretty exceptions when caught by the main REPL loop. Also, overrides
   `clojure.repl/pst`, `clojure.stacktrace/print-stack-trace`, `clojure.stacktrace/print-cause-trace`.
+
+  In addition, installs an [[uncaught-exception-handler]] so that uncaught exceptions in non-REPL threads
+  will be printed reasonably. See [[io.aviso.logging]] for a better handler, used when clojure.tools.logging
+  is available.
 
   Caught exceptions do not print the stack trace; the pst replacement does."
   []
@@ -76,4 +69,59 @@
   (reset-var! #'repl/pst pretty-pst)
   (reset-var! #'st/print-stack-trace pretty-print-stack-trace)
   (reset-var! #'st/print-cause-trace pretty-print-stack-trace)
+
+  ;; This is necessary for Clojure 1.8 and above, due to direct linking
+  ;; (from clojure.test to clojure.stacktrace).
+  (RT/loadResourceScript "clojure/test.clj")
+
+  (Thread/setDefaultUncaughtExceptionHandler (uncaught-exception-handler))
   nil)
+
+(defn ^String copy
+  "Copies the current contents of the Clipboard, returning its contents as a string.
+
+  This makes use of AWT; it will throw java.awt.HeadlessException when AWT is not
+  available, for example, when the JVM is launched with `-Djava.awt.headless=true`."
+  {:added "0.1.32"}
+  []
+  (require 'boot.from.io.aviso.clipboard)
+  ((ns-resolve 'boot.from.io.aviso.clipboard 'copy)))
+
+(defn pretty-print
+  "Pretty-prints the supplied object to a returned string.
+
+  With no arguments, copies from the clipboard, parses as EDN, and prints the EDN data to `*out*`,
+  returning nil."
+  {:added "0.1.32"}
+  ([]
+   (-> (copy) edn/read-string pprint))
+  ([object]
+   (write object
+          :stream nil
+          :pretty true)))
+
+(defn paste
+  "Pastes a string in as the new content of the Clipboard.
+
+  This can be helpful when, for example, pretty printing some EDN content from a log file
+  before pasting it into some other editor."
+  {:added "0.1.32"}
+  [^String s]
+  (require 'boot.from.io.aviso.clipboard)
+  ((ns-resolve 'boot.from.io.aviso.clipboard 'paste) s))
+
+(defn format-exception
+  "Passed the standard exception text and formats it using [[parse-exception]] and
+  [[write-exception]], returning the formatted exception text.
+
+  With no arguments, parses the clipboard text and prints the formatted exception
+  to `*out*` (returning nil)."
+  {:added "0.1.32"}
+  ([]
+   (e/write-exception* *out*
+                       (-> (copy)
+                           (e/parse-exception nil))
+                       nil))
+  ([text]
+   (writer/into-string
+     e/write-exception* (e/parse-exception text nil) nil)))